+++ Microsoft.Azure.Cosmos/src/Resource/Container/ContainerResponse.cs
@@ -28,12 +28,14 @@ namespace Microsoft.Azure.Cosmos
             HttpStatusCode httpStatusCode,
             Headers headers,
             ContainerProperties containerProperties,
-            Container container)
+            Container container,
+            CosmosDiagnostics diagnostics)
         {
             this.StatusCode = httpStatusCode;
             this.Headers = headers;
             this.Resource = containerProperties;
             this.Container = container;
+            this.Diagnostics = diagnostics;
         }
 
         /// <summary>
@@ -50,6 +52,9 @@ namespace Microsoft.Azure.Cosmos
 
         /// <inheritdoc/>
         public override HttpStatusCode StatusCode { get; }
+
+        /// <inheritdoc/>
+        public override CosmosDiagnostics Diagnostics { get; }
 
         /// <inheritdoc/>
         public override double RequestCharge => this.Headers?.RequestCharge ?? 0;